+++ src/tty/windows.rs
@@ -54,13 +54,13 @@ fn check(rc: BOOL) -> io::Result<()> {
     }
 }
 
-fn get_win_size(handle: HANDLE) -> (usize, usize) {
+fn get_win_size(handle: HANDLE) -> (u16, u16) {
     let mut info = unsafe { mem::zeroed() };
     match unsafe { wincon::GetConsoleScreenBufferInfo(handle, &mut info) } {
         FALSE => (80, 24),
         _ => (
-            info.dwSize.X as usize,
-            (1 + info.srWindow.Bottom - info.srWindow.Top) as usize,
+            info.dwSize.X as u16,
+            (1 + info.srWindow.Bottom - info.srWindow.Top) as u16,
         ), // (info.srWindow.Right - info.srWindow.Left + 1)
     }
 }
@@ -286,7 +286,7 @@ fn read_input(handle: HANDLE, max_count: u32) -> Result<KeyEvent> {
 
 pub struct ConsoleRenderer {
     conout: HANDLE,
-    cols: usize, // Number of columns in terminal
+    cols: u16, // Number of columns in terminal
     buffer: String,
     utf16: Vec<u16>,
     colors_enabled: bool,
@@ -526,13 +526,13 @@ impl Renderer for ConsoleRenderer {
         self.cols = cols;
     }
 
-    fn get_columns(&self) -> usize {
+    fn get_columns(&self) -> u16 {
         self.cols
     }
 
     /// Try to get the number of rows in the current terminal,
     /// or assume 24 if it fails.
-    fn get_rows(&self) -> usize {
+    fn get_rows(&self) -> u16 {
         let (_, rows) = get_win_size(self.conout);
         rows
     }
@@ -641,7 +641,7 @@ impl Term for Console {
     fn new(
         color_mode: ColorMode,
         behavior: Behavior,
-        _tab_stop: usize,
+        _tab_stop: u16,
         bell_style: BellStyle,
         _enable_bracketed_paste: bool,
     ) -> Console {
